+++ SHELL/SCRIPTS/test_scenario.py
@@ -1,0 +1,7 @@
+from SHELL.TestShellApplication import *
+
+
+class TestScenario(ABC):
+    @abstractmethod
+    def run_test(self, shell: TestShellApplication):
+        pass
